Successfully querying for EMS room availability and converting data to a struct
