+++ helpers/availability.go
@@ -1,6 +1,26 @@
 package helpers
 
-import "fmt"
+import (
+	"encoding/xml"
+	"fmt"
+	"os"
+	"time"
+)
+
+type RoomAvailabilityRequest struct {
+	XMLName     struct{} `xml:"http://DEA.EMS.API.Web.Service/ GetRoomAvailability"`
+	Username    string   `xml:"UserName"`
+	Password    string
+	RoomID      int
+	BookingDate time.Time
+	StartTime   time.Time
+	EndTime     time.Time
+}
+
+type RoomAvailabilityResponse struct {
+	XMLName struct{} `xml:"http://DEA.EMS.API.Web.Service/ GetRoomAvailabilityResponse"`
+	Result  string   `xml:"GetRoomAvailabilityResult"`
+}
 
 // CheckAvailability checks room availability by consulting with the EMS API and trying to ping the room via telnet
 func CheckAvailability(building string, room string) bool {
@@ -21,12 +41,31 @@ func CheckTelnetAvailability() bool {
 
 // CheckEMSAvailability consults the EMS API to see if the room in question is scheduled to be in use currently
 func CheckEMSAvailability(building string, room string) bool {
-	availability, err := GetRoomID(building, room)
+	roomID, err := GetRoomID(building, room)
 	CheckErr(err)
 
-	fmt.Printf("%v", availability)
+	fmt.Printf("%v", roomID)
 
-	// TODO Check to availability after obtaining room ID
+	date := time.Now()
+	startTime := time.Now()
+	endTime := time.Now()
+
+	fmt.Printf("Date: %v, Start: %v, End: %v\n", date, startTime, endTime)
+
+	request := &RoomAvailabilityRequest{Username: os.Getenv("EMS_API_USERNAME"), Password: os.Getenv("EMS_API_PASSWORD"), RoomID: roomID, BookingDate: date, StartTime: startTime, EndTime: endTime}
+	encodedRequest, err := SoapEncode(&request)
+	CheckErr(err)
+
+	response := SoapRequest("https://emsweb-dev.byu.edu/EMSAPI/Service.asmx", encodedRequest)
+	availability := RoomAvailabilityResponse{}
+	err = SoapDecode([]byte(response), &availability)
+	CheckErr(err)
+
+	roomAvailability := Room{}
+	err = xml.Unmarshal([]byte(availability.Result), &roomAvailability)
+	CheckErr(err)
+
+	fmt.Printf("%v", roomAvailability.Available)
 
 	return true
 }
